ref(feedback): Colocate feedback types back into the packages/feedback/ folder
