+++ packages/feedback/src/modal/components/Dialog.tsx
@@ -1,9 +1,9 @@
-import type { FeedbackFormData } from '@sentry/types';
 // biome-ignore lint/nursery/noUnusedImports: reason
 import { Fragment, h } from 'preact'; // eslint-disable-line @typescript-eslint/no-unused-vars
 import type { VNode } from 'preact';
 import { useCallback, useMemo, useState } from 'preact/hooks';
 import { SUCCESS_MESSAGE_TIMEOUT } from '../../constants';
+import type { FeedbackFormData } from '../../types';
 import { DialogHeader } from './DialogHeader';
 import type { Props as HeaderProps } from './DialogHeader';
 import type { Props as FormProps } from './Form';
