+++ packages/feedback/src/modal/components/DialogHeader.tsx
@@ -1,8 +1,8 @@
-import type { FeedbackInternalOptions } from '@sentry/types';
 // biome-ignore lint/nursery/noUnusedImports: reason
 import { h } from 'preact'; // eslint-disable-line @typescript-eslint/no-unused-vars
 import type { VNode } from 'preact';
 import { useMemo } from 'preact/hooks';
+import type { FeedbackInternalOptions } from '../../types';
 import { SentryLogo } from './SentryLogo';
 
 export interface Props {
