+++ packages/feedback/src/screenshot/components/ScreenshotEditor.tsx
@@ -1,10 +1,10 @@
-import type { FeedbackDialog } from '@sentry/types';
 /* eslint-disable max-lines */
 import type { ComponentType, VNode, h as hType } from 'preact';
 // biome-ignore lint: needed for preact
 import { h } from 'preact'; // eslint-disable-line @typescript-eslint/no-unused-vars
 import { useCallback, useEffect, useMemo, useRef, useState } from 'preact/hooks';
 import { CROP_COLOR, DOCUMENT, WINDOW } from '../../constants';
+import type { FeedbackDialog } from '../../types';
 import { createScreenshotInputStyles } from './ScreenshotInput.css';
 import { useTakeScreenshot } from './useTakeScreenshot';
