+++ packages/feedback/src/screenshot/integration.ts
@@ -1,7 +1,7 @@
-import type { FeedbackDialog, FeedbackScreenshotIntegration, IntegrationFn } from '@sentry/types';
-import type { Attachment } from '@sentry/types';
+import type { Attachment, IntegrationFn } from '@sentry/types';
 import type { h as hType } from 'preact';
 import { DOCUMENT } from '../constants';
+import type { FeedbackDialog, FeedbackScreenshotIntegration } from '../types';
 import { makeScreenshotEditorComponent } from './components/ScreenshotEditor';
 
 export const feedbackScreenshotIntegration = ((): FeedbackScreenshotIntegration => {
