+++ packages/feedback/src/util/validate.ts
@@ -1,4 +1,4 @@
-import type { FeedbackFormData, FeedbackInternalOptions } from '@sentry/types';
+import type { FeedbackFormData, FeedbackInternalOptions } from '../types';
 
 export type Props = Pick<
   FeedbackInternalOptions,
